app/requirements.txt
RPi.GPIO
spidev
paho-mqtt